+++ protocol/src/main/java/org/geysermc/mcprotocollib/network/packet/PacketRegistry.java
@@ -6,6 +6,10 @@ import it.unimi.dsi.fastutil.ints.Int2ObjectOpenHashMap;
 import org.geysermc.mcprotocollib.network.codec.PacketCodecHelper;
 import org.geysermc.mcprotocollib.network.codec.PacketDefinition;
 import org.geysermc.mcprotocollib.network.codec.PacketSerializer;
+import org.geysermc.mcprotocollib.protocol.codec.MinecraftCodecHelper;
+import org.geysermc.mcprotocollib.protocol.codec.MinecraftPacket;
+import org.geysermc.mcprotocollib.protocol.codec.MinecraftPacketSerializer;
+import org.geysermc.mcprotocollib.protocol.codec.PacketFactory;
 
 import java.util.IdentityHashMap;
 import java.util.Map;
@@ -246,5 +250,52 @@ public class PacketRegistry {
         }
 
         return definition;
+    }
+
+    public static Builder builder() {
+        return new Builder();
+    }
+
+    public static class Builder {
+        private final Int2ObjectMap<PacketDefinition<? extends MinecraftPacket, MinecraftCodecHelper>> clientboundPackets = new Int2ObjectOpenHashMap<>();
+        private final Int2ObjectMap<PacketDefinition<? extends MinecraftPacket, MinecraftCodecHelper>> serverboundPackets = new Int2ObjectOpenHashMap<>();
+
+        private int nextClientboundId = 0x00;
+        private int nextServerboundId = 0x00;
+
+        public <T extends MinecraftPacket> Builder registerClientboundPacket(int id, Class<T> packetClass, PacketFactory<T, MinecraftCodecHelper> factory) {
+            this.nextClientboundId = id;
+            return registerClientboundPacket(packetClass, factory);
+        }
+
+        public <T extends MinecraftPacket> Builder registerClientboundPacket(Class<T> packetClass, PacketFactory<T, MinecraftCodecHelper> factory) {
+            this.clientboundPackets.put(nextClientboundId, new PacketDefinition<>(nextClientboundId, packetClass, new MinecraftPacketSerializer<>(factory)));
+            this.nextClientboundId++;
+            return this;
+        }
+
+        public <T extends MinecraftPacket> Builder registerServerboundPacket(int id, Class<T> packetClass, PacketFactory<T, MinecraftCodecHelper> factory) {
+            this.nextServerboundId = id;
+            return registerServerboundPacket(packetClass, factory);
+        }
+
+        public <T extends MinecraftPacket> Builder registerServerboundPacket(Class<T> packetClass, PacketFactory<T, MinecraftCodecHelper> factory) {
+            this.serverboundPackets.put(nextServerboundId, new PacketDefinition<>(nextServerboundId, packetClass, new MinecraftPacketSerializer<>(factory)));
+            this.nextServerboundId++;
+            return this;
+        }
+
+        public PacketRegistry build() {
+            PacketRegistry codec = new PacketRegistry();
+            for (Int2ObjectMap.Entry<PacketDefinition<? extends MinecraftPacket, MinecraftCodecHelper>> entry : this.clientboundPackets.int2ObjectEntrySet()) {
+                codec.registerClientbound(entry.getValue());
+            }
+
+            for (Int2ObjectMap.Entry<PacketDefinition<? extends MinecraftPacket, MinecraftCodecHelper>> entry : this.serverboundPackets.int2ObjectEntrySet()) {
+                codec.registerServerbound(entry.getValue());
+            }
+
+            return codec;
+        }
     }
 }
